Rename itens incorretos
Renomeei alguns itens.
