+++ CsvScreen/pom.xml
@@ -5,7 +5,7 @@
     <modelVersion>4.0.0</modelVersion>
 
     <groupId>com.example</groupId>
-    <artifactId>seila</artifactId>
+    <artifactId>CsvScreen</artifactId>
     <version>1.0-SNAPSHOT</version>
     <name>seila</name>
 
@@ -60,7 +60,7 @@
                         <!-- Default configuration for running with: mvn clean javafx:run -->
                         <id>default-cli</id>
                         <configuration>
-                            <mainClass>com.example.seila/com.example.seila.HelloApplication</mainClass>
+                            <mainClass>com.example.seila/com.example.csvscreen.HelloApplication</mainClass>
                             <launcher>app</launcher>
                             <jlinkZipName>app</jlinkZipName>
                             <jlinkImageName>app</jlinkImageName>
